e: massiv warning cleanup !
SVN revision: 70756

+++ src/modules/everything/evry_api.h
@@ -252,7 +252,7 @@ struct _Evry_Event_Action_Performed
 
 #define EVRY_PLUGIN_MIN_QUERY(_p, _input)				\
   if (!(EVRY_PLUGIN(_p)->config->min_query) ||				\
-      (_input && (strlen(_input) >= EVRY_PLUGIN(_p)->config->min_query)))
+      (_input && ((int) strlen(_input) >= EVRY_PLUGIN(_p)->config->min_query)))
 
 #define EVRY_PLUGIN_ITEMS_CLEAR(_p) {				\
      Evry_Item *_it;						\
